cleaned up User Interactors

+++ src/main/java/com/newts/newtapp/api/application/user/DeleteUser.java
@@ -2,6 +2,8 @@ package com.newts.newtapp.api.application.user;
 
 import com.newts.newtapp.api.UserRepository;
 import com.newts.newtapp.api.application.*;
+import com.newts.newtapp.api.errors.IncorrectPassword;
+import com.newts.newtapp.api.errors.InvalidPassword;
 import com.newts.newtapp.api.errors.UserNotFound;
 import com.newts.newtapp.entities.User;
 
@@ -19,12 +21,16 @@ public class DeleteUser extends UserInteractor<Void,Exception> {
      * @param request   a request stored as a RequestModel
      */
     @Override
-    public Void request(RequestModel request) throws UserNotFound {
+    public Void request(RequestModel request) throws UserNotFound, IncorrectPassword {
         int userId = (int) request.get(RequestField.ID);
-
+        String password = (String) request.get(RequestField.PASSWORD);
         User user = repository.findById(userId).orElseThrow(UserNotFound::new);
-        repository.delete(user);
 
+        if (!(user.getPassword().equals(password))) {
+            throw new IncorrectPassword();
+        }
+
+        repository.delete(user);
         return null;
        }
 }
